fix path and move index into pages
