+++ pages/account.php
@@ -27,7 +27,7 @@
 </head>
 <body>
     <?php 
-        require_once "./pages/Components/header.php";
+        require_once "./Components/header.php";
     ?>
 
     <section id="profile">
@@ -171,7 +171,7 @@
     </section>
 
     <?php
-        require_once "./pages/Components/footer.php";
+        require_once "./Components/footer.php";
     ?>
 </body>
 </html>
