+++ pages/explore.php
@@ -13,7 +13,7 @@
         { 
             session_start(); 
         } 
-        require_once "./pages/Components/header.php";
+        require_once "./Components/header.php";
         $base_url = 'http://' . $_SERVER['HTTP_HOST'] . rtrim(dirname(dirname($_SERVER['PHP_SELF'])), '/\\');
         if(isset($_SESSION['Student_status']) && $_SESSION['Student_status'] == 'banned') {
             header("Location: $base_url/pages/index.php?page=you_have_been_banned");
@@ -211,7 +211,7 @@
     
 
     <?php
-        require_once "./pages/Components/footer.php";
+        require_once "./Components/footer.php";
     ?>
 
     <!-- <script src="../js/Page/explore.js"></script> -->
