+++ pages/question_edit.php
@@ -10,7 +10,7 @@
 <body>
     <?php 
         $base_url = 'http://' . $_SERVER['HTTP_HOST'] . rtrim(dirname(dirname($_SERVER['PHP_SELF'])), '/\\');
-        require_once "./pages/Components/header.php";
+        require_once "./Components/header.php";
         require_once "../logical/database_connect.php";
         require_once "../logical/function.php";
         $question_id = sanitize_input($_GET['question_id']);
@@ -138,7 +138,7 @@
    <script src="../js/Page/add_question.js"></script>
 
     <?php
-        include "./pages/Components/footer.php";
+        include "./Components/footer.php";
     ?>
 </body>
 </html>
