+++ pages/reset_password.php
@@ -53,7 +53,7 @@
     </section> -->
 
     <?php
-        require_once "./pages/Components/footer.php";
+        require_once "./Components/footer.php";
     ?>
     <script type="module" src="../js/Page/reset_password.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
