+++ pages/sign_up.php
@@ -66,7 +66,7 @@ if(!isset($_SESSION))
     <script type="module" src="../js/Page/sign_up.js"></script>
 
     <?php
-        require_once "./pages/Components/footer.php"
+        require_once "./Components/footer.php"
     ?>
 </body>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
